refactor: MultilineTextViewer 상세 페이지 관련 컴포넌트로 폴더 위치 변경
