+++ frontend/src/pages/DetailedReviewPage/components/MultilineTextViewer/styles.ts
