+++ frontend/src/pages/DetailedReviewPage/components/index.tsx
@@ -1,4 +1,6 @@
-export { default as MultipleChoiceAnswerList } from './MultipleChoiceAnswerList';
-export { default as ReviewDescription } from './ReviewDescription';
 export { default as QuestionAnswerSection } from './QuestionAnswerSection';
+export { default as QuestionTitle } from './QuestionTitle';
+export { default as MultipleChoiceAnswerList } from './MultipleChoiceAnswerList';
+export { default as MultilineTextViewer } from './MultilineTextViewer';
+export { default as ReviewDescription } from './ReviewDescription';
 export { default as DetailedReviewPageContents } from './DetailedReviewPageContents';
